feat: clean authz test
